dashboard/pyproject.toml
[tool.poetry]
name = "monster-dashboard"
version = "0.1.0"
description = ""
authors = ["Scott Persinger <scottpersinger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
reflex = "0.5.9"
pydantic = "1.10.14"
html2text = "^2020.1.16"
google-auth = {extras = ["requests"], version = "^2.27.0"}
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"
pyyaml = "^6.0.1"
openai = "^1.12.0"
pillow = "^10.2.0"
requests-oauthlib = "^1.3.1"
python-dotenv = "^1.0.1"
timeago = "^1.0.16"
apscheduler = "^3.10.4"
openapi = "^1.1.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
markdown2 = "^2.4.13"
bleach = "^6.1.0"
boto3 = "^1.34.69"
dill = "^0.3.8"
aiofiles = "^23.2.1"
arrow = "^1.3.0"
markdown = "^3.6"
pygments = "^2.18.0"
rollbar = "^1.0.0"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"
pypdf = "^4.3.1"
pypdf2 = "^3.0.1"
reflex-icons = "^1.0.2"
fastapi-lifespan-manager = "^0.1.4"
slack-bolt = "^1.20.1"
aiohttp = "^3.10.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
